add functionality to delete images from aws when restaurant is deleted

+++ src/restaurants/restaurants.controller.ts
@@ -56,12 +56,19 @@ export class RestaurantsController {
   async deleteRestaurant(
     @Param('id') id: string,
   ): Promise<{ deleted: Boolean }> {
-    await this.restaurantsService.findById(id);
+    const restaurant = await this.restaurantsService.findById(id);
+    const isDeleted = await this.restaurantsService.deleteImages(
+      restaurant.images,
+    );
 
-    const restaurant = this.restaurantsService.deleteById(id);
-    if (restaurant) {
+    if (isDeleted) {
+      this.restaurantsService.deleteById(id);
       return {
         deleted: true,
+      };
+    } else {
+      return {
+        deleted: false,
       };
     }
   }
